#2395 delete wp buttons
